options: switch storage management to message passing (#30)
* options: delete unused whitelist variable

Nothing in this page uses this variable, so stop loading it.

* options: switch storage management to message passing

Using chrome.runtime.getBackgroundPage to get the background page object
and then accessing the storage state directly has always been a bit of a
hack.  This API has been dropped from MV3 entirely.  Switch to message
passing like many of the other options already to be MV3 compatible.

---------

Co-authored-by: Rob Kodey <rkodey+github@gmail.com>